Pupils not taking the check.
