+++ admin/routes/school.js
@@ -9,13 +9,9 @@ const { getHome,
   postSubmitAttendance,
   getDeclarationForm,
   postDeclarationForm,
-  getHDFSubmitted,
-  getPupilNotTakingCheck,
-  getSelectPupilNotTakingCheck,
-  savePupilNotTakingCheck,
-  removePupilNotTakingCheck,
-  viewPupilsNotTakingTheCheck } = require('../controllers/school')
+  getHDFSubmitted } = require('../controllers/school')
 const group = require('../controllers/group.js')
+const pupilsNotTakingTheCheck = require('../controllers/pupils-not-taking-the-check')
 
 const school = (router) => {
   router.get('/school-home', isAuthenticated(rolesConfig.ROLE_TEACHER), (req, res, next) => getHome(req, res, next))
@@ -27,14 +23,14 @@ const school = (router) => {
   router.get('/declaration-form', isAuthenticated(rolesConfig.ROLE_TEACHER), (req, res, next) => getDeclarationForm(req, res, next))
   router.post('/submit-declaration-form', isAuthenticated(rolesConfig.ROLE_TEACHER), (req, res, next) => postDeclarationForm(req, res, next))
   router.get('/declaration-form-submitted', isAuthenticated(rolesConfig.ROLE_TEACHER), (req, res, next) => getHDFSubmitted(req, res, next))
-  router.get('/pupils-not-taking-check/select-pupils/:groupIds?', isAuthenticated(rolesConfig.ROLE_TEACHER), (req, res, next) => getSelectPupilNotTakingCheck(req, res, next))
-  router.get('/pupils-not-taking-check/select-pupils/:sortField/:sortDirection/:groupIds?', isAuthenticated(rolesConfig.ROLE_TEACHER), (req, res, next) => getSelectPupilNotTakingCheck(req, res, next))
-  router.get('/pupils-not-taking-check/save-pupils', isAuthenticated(rolesConfig.ROLE_TEACHER), (req, res, next) => getSelectPupilNotTakingCheck(req, res, next))
-  router.post('/pupils-not-taking-check/save-pupils', isAuthenticated(rolesConfig.ROLE_TEACHER), (req, res, next) => savePupilNotTakingCheck(req, res, next))
-  router.get('/pupils-not-taking-check/remove/:pupilId', isAuthenticated(rolesConfig.ROLE_TEACHER), (req, res, next) => removePupilNotTakingCheck(req, res, next))
-  router.get('/pupils-not-taking-check/view', isAuthenticated(rolesConfig.ROLE_TEACHER), (req, res, next) => viewPupilsNotTakingTheCheck(req, res, next))
-  router.get('/pupils-not-taking-check', isAuthenticated(rolesConfig.ROLE_TEACHER), (req, res, next) => getPupilNotTakingCheck(req, res, next))
-  router.get('/pupils-not-taking-check/:removed', isAuthenticated(rolesConfig.ROLE_TEACHER), (req, res, next) => getPupilNotTakingCheck(req, res, next))
+  router.get('/pupils-not-taking-check/select-pupils/:groupIds?', isAuthenticated(rolesConfig.ROLE_TEACHER), (req, res, next) => pupilsNotTakingTheCheck.getSelectPupilNotTakingCheck(req, res, next))
+  router.get('/pupils-not-taking-check/select-pupils/:sortField/:sortDirection/:groupIds?', isAuthenticated(rolesConfig.ROLE_TEACHER), (req, res, next) => pupilsNotTakingTheCheck.getSelectPupilNotTakingCheck(req, res, next))
+  router.get('/pupils-not-taking-check/save-pupils', isAuthenticated(rolesConfig.ROLE_TEACHER), (req, res, next) => pupilsNotTakingTheCheck.getSelectPupilNotTakingCheck(req, res, next))
+  router.post('/pupils-not-taking-check/save-pupils', isAuthenticated(rolesConfig.ROLE_TEACHER), (req, res, next) => pupilsNotTakingTheCheck.savePupilNotTakingCheck(req, res, next))
+  router.get('/pupils-not-taking-check/remove/:pupilId', isAuthenticated(rolesConfig.ROLE_TEACHER), (req, res, next) => pupilsNotTakingTheCheck.removePupilNotTakingCheck(req, res, next))
+  router.get('/pupils-not-taking-check/view', isAuthenticated(rolesConfig.ROLE_TEACHER), (req, res, next) => pupilsNotTakingTheCheck.viewPupilsNotTakingTheCheck(req, res, next))
+  router.get('/pupils-not-taking-check', isAuthenticated(rolesConfig.ROLE_TEACHER), (req, res, next) => pupilsNotTakingTheCheck.getPupilNotTakingCheck(req, res, next))
+  router.get('/pupils-not-taking-check/:removed', isAuthenticated(rolesConfig.ROLE_TEACHER), (req, res, next) => pupilsNotTakingTheCheck.getPupilNotTakingCheck(req, res, next))
   router.get('/group-pupils', isAuthenticated(rolesConfig.ROLE_TEACHER), (req, res, next) => group.groupPupilsPage(req, res, next))
   router.get('/group-pupils/add', isAuthenticated(rolesConfig.ROLE_TEACHER), (req, res, next) => group.manageGroupPage(req, res, next))
   router.get('/group-pupils/edit/:groupId', isAuthenticated(rolesConfig.ROLE_TEACHER), (req, res, next) => group.manageGroupPage(req, res, next))
