+++ admin/views/school/select-pupils-not-taking-check.ejs
@@ -108,7 +108,7 @@
                             </label>
                             <% if (pupil.showDoB) { %>
                             <div class="font-greyed-out font-xsmall">
-                                Date of birth: <%= pupil.dob %>
+                                Date of birth: <%= pupil.dateOfBirth %>
                             </div>
                             <% } %>
                         </td>
